+++ common/src/main/java/com/walmartlabs/concord/common/ReflectionUtils.java
@@ -1,0 +1,50 @@
+package com.walmartlabs.concord.common;
+
+/*-
+ * *****
+ * Concord
+ * -----
+ * Copyright (C) 2017 - 2018 Walmart Inc.
+ * -----
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * 
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ * 
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * =====
+ */
+
+import java.lang.annotation.Annotation;
+
+public final class ReflectionUtils {
+
+    public static <A extends Annotation> A findAnnotation(Class<?> clazz, Class<A> annotationType) {
+        A annotation = clazz.getAnnotation(annotationType);
+        if (annotation != null) {
+            return annotation;
+        }
+
+        for (Class<?> ifc : clazz.getInterfaces()) {
+            annotation = findAnnotation(ifc, annotationType);
+            if (annotation != null) {
+                return annotation;
+            }
+        }
+
+        Class<?> superClass = clazz.getSuperclass();
+        if (superClass == null || superClass == Object.class) {
+            return null;
+        }
+
+        return findAnnotation(superClass, annotationType);
+    }
+
+    private ReflectionUtils() {
+    }
+}