+++ project-model/src/main/java/com/walmartlabs/concord/project/yaml/converter/StepConverter.java
@@ -159,7 +159,6 @@ public interface StepConverter<T extends YamlStep> {
         if (key.equals("in") && getWithItems(options) != null) {
            result = appendWithItemsVar(result);
         }
-
         return result;
     }